+++ autoload.php
@@ -5,17 +5,17 @@
  */
 
 // Register autoload function
-spl_autoload_register(function (string $class) {
+\spl_autoload_register(function (string $class) {
     // Only autoload classes shipped with the plugin.
-    if (strpos($class, 'BlueChip\\Cache') !== 0) {
+    if (!\str_starts_with($class, 'BlueChip\\Cache')) {
         return;
     }
 
     // Get absolute name of class file
-    $file = __DIR__ . '/classes/' . str_replace('\\', '/', $class) . '.php';
+    $file = __DIR__ . '/classes/' . \str_replace('\\', '/', $class) . '.php';
 
     // If the class file is readable, load it!
-    if (is_readable($file)) {
+    if (\is_readable($file)) {
         require_once $file;
     }
 });