+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # BC Cache Changelog
 
+## Upcoming version 3.1.0 (2023-01-04)
+
+### Changed
+
+* WordPress 5.9 or newer is now required [#88](https://github.com/chesio/bc-cache/issues/88).
+* Improved serialization of data stored in database [#90](https://github.com/chesio/bc-cache/issues/90).
+* On WordPress 6.1 and newer, output buffering is started in `send_headers` action already [#87](https://github.com/chesio/bc-cache/issues/87).
+
+### Fixed
+
+* Prevent deprecation notices on PHP 8.1 and newer [#89](https://github.com/chesio/bc-cache/issues/89).
+
 ## Version 3.0.1 (2022-12-28)
 
 ### Fixed